refactor(lotto): rename Result to Statistic
